requirements.txt
streamlit 
openai 
gtts 
pypdf
pandas
python-dotenv
langchain
langchain_community
langchain-openai
xlrd